requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
apify_client==1.9.2
apify_shared==1.3.2
asgiref==3.8.1
attrs==24.3.0
beautifulsoup4==4.14.3
billiard==4.2.1
bleach==6.3.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.42.2
colorama==0.4.6
deprecation==2.1.0
dj-database-url==2.3.0
Django==5.1.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-filter==24.3
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
ffmpeg==1.4
frozenlist==1.5.0
gotrue==2.11.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2025.4.15
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
inflection==0.5.1
kombu==5.4.2
mailchimp-marketing==3.0.80
mailgun==0.1.1
more-itertools==10.6.0
multidict==6.1.0
mysqlclient==2.2.6
packaging==24.2
pillow==11.0.0
postgrest==0.19.1
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg2==2.9.10
pyactiveresource==1.0.1
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2024.2
PyYAML==6.0.2
realtime==2.1.0
redis==7.1.0
requests==2.32.3
sib-api-v3-sdk==7.6.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sqlparse==0.5.1
storage3==0.11.0
StrEnum==0.4.15
supabase==2.11.0
supafunc==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==13.1
whitenoise==6.8.2
yarl==1.18.3
zipp==3.21.0
